+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `logs.FORMAT_SEC`, rename `FORMAT` -> `FORMAT_MS`.
+- New class `logs.LoggingSetupHelper`.
+- New functions `logs.get_logger()` and `convert_log_level()`.
+
+### Changed
+- Update `basic_config.basic_config()`: Allow and handle `level=None` to avoid logging from root.
+
 ## [5.0.1] - 2025-02-01
 
 ### Fixed